[FEAT] Create Create Challenge & Participate Challenge Logic

+++ src/main/java/kb/wgwg/domain/CoffeeChallenge.java
@@ -22,11 +22,11 @@ import java.time.LocalDateTime;
 @DiscriminatorValue(value = "COFFEE")
 public class CoffeeChallenge extends Challenge {
 
-    private BigDecimal totalAsset;
+    private int totalAsset;
 
     @Builder
     public CoffeeChallenge(String title, String description, String status,
-                           LocalDateTime startDate, LocalDateTime endDate, BigDecimal totalAsset)
+                           LocalDateTime startDate, LocalDateTime endDate, int totalAsset)
     {
         super(title, description, status, startDate, endDate);
         this.totalAsset = totalAsset;
